Refactor: Adicionando A interface service, antes da serviceImpl e ajustando alguns problemas do merge
