+++ backend-rupay/src/main/resources/application.properties
@@ -6,7 +6,7 @@ spring.datasource.username=postgres
 spring.datasource.password=admin
 
 spring.jpa.properties..hibernate.dialect = org.hibernate.dialect.PostgreSQLDialect
-spring.jpa.hibernate.ddl-auto=create-drop
+spring.jpa.hibernate.ddl-auto=update
 spring.jpa.show-sql=true
 spring.jpa.properties.hibernate.format_sql=true
